refacor(code-block): some small tweaks, and add a placeholder style sheet
